requirements.txt
opencv-python>=4.8.0
numpy>=1.21.2
pyyaml>=6.0
imutils>=0.5.4
pillow>=10.0
ultralytics>=8.0.0